[#259] Extract HTTP server into service base.
Extracted HTTP server functionaliy from AbstractVertxBasedHttpProtocolAdapter.java
into new class AbstractVertxHttpServer in hono-service-base.
